Changes in master blade of session and in app.js
